+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 ### Feature
 * Update package to be compatible with v5.0.0
 * Added tools.RestoreIntents which will restore Intents, Groups, & Dashboard from a file, another instance, or factory default
+* Added all inventory endpoints to inventory (phones, fans, modules, powerSupplies, powerSuppliesFans)
+* Allow initialization without loaded snapshots
+* Added IPFClient.loaded_snapshots and IPFClient.unloaded_snapshots
+* Added a Load and Unload snapshot function to Snapshot model `ipf.snapshots['ea0da479-6222-4cf6-a0ae-78a610669477'].load(ipf)`
 
 ## 4.4.3 (2022-06-22)
 ### Feature